+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ByteriX, 2017. All right reserved.
 ##### Improvements
 * documentation for action, date, pictures, rate, suggestions, variants, text row.
 * move & rename all picture items
+* extract item classes from suggestion & variants rows
 
 ## 0.9.37 (15.02.2017)
 ##### Improvements